+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ## Unreleased
-- parser: Added support for OGNLT24 (LT24), OGSKYL (Skylines), OGSPID (Spider) and OGSPOT (Spot)
+- parser: Added support for OGNLT24 (LT24), OGSKYL (Skylines), OGSPID (Spider), OGSPOT (Spot) and OGNFNT (Fanet)
+- parser: Added support for (server) comments
 - parser: Added parser for local receiver output (port 50001)
 
 ## 0.8.2: - 2018-01-20